Moving MultiBrokerMetricsTest out from MetricsTest
